Tony refactor routing (#62)
