+++ src/main/kotlin/org/bibletranslationtools/fetcher/web/controllers/HomeController.kt
@@ -1,0 +1,28 @@
+package org.bibletranslationtools.fetcher.web.controllers
+
+import dev.jbs.ktor.thymeleaf.ThymeleafContent
+import io.ktor.application.call
+import io.ktor.response.respond
+import io.ktor.routing.Routing
+import io.ktor.routing.get
+import io.ktor.routing.route
+import org.bibletranslationtools.fetcher.web.controllers.utils.GL_ROUTE
+import org.bibletranslationtools.fetcher.web.controllers.utils.contentLanguage
+import org.bibletranslationtools.fetcher.web.controllers.utils.getPreferredLocale
+
+fun Routing.homeController() {
+    route("/") {
+        get {
+            // landing page
+            call.respond(
+                ThymeleafContent(
+                    template = "landing",
+                    model = mapOf(
+                        "glRoute" to "/$GL_ROUTE"
+                    ),
+                    locale = getPreferredLocale(contentLanguage, "landing")
+                )
+            )
+        }
+    }
+}
