+++ src/main/kotlin/org/bibletranslationtools/fetcher/web/controllers/ProductController.kt
@@ -1,0 +1,70 @@
+package org.bibletranslationtools.fetcher.web.controllers
+
+import dev.jbs.ktor.thymeleaf.ThymeleafContent
+import io.ktor.application.call
+import io.ktor.http.HttpStatusCode
+import io.ktor.request.path
+import io.ktor.response.respond
+import io.ktor.routing.Routing
+import io.ktor.routing.get
+import io.ktor.routing.route
+import java.util.Locale
+import org.bibletranslationtools.fetcher.usecase.DependencyResolver
+import org.bibletranslationtools.fetcher.usecase.FetchProductViewData
+import org.bibletranslationtools.fetcher.web.controllers.utils.GL_ROUTE
+import org.bibletranslationtools.fetcher.web.controllers.utils.LANGUAGE_PARAM_KEY
+import org.bibletranslationtools.fetcher.web.controllers.utils.RoutingValidator
+import org.bibletranslationtools.fetcher.web.controllers.utils.UrlParameters
+import org.bibletranslationtools.fetcher.web.controllers.utils.contentLanguage
+import org.bibletranslationtools.fetcher.web.controllers.utils.errorPage
+import org.bibletranslationtools.fetcher.web.controllers.utils.getLanguageName
+import org.bibletranslationtools.fetcher.web.controllers.utils.getPreferredLocale
+import org.bibletranslationtools.fetcher.web.controllers.utils.normalizeUrl
+
+fun Routing.productController(resolver: DependencyResolver) {
+    route("/$GL_ROUTE/{$LANGUAGE_PARAM_KEY}") {
+        get {
+            // products page
+            val path = normalizeUrl(call.request.path())
+            val params = UrlParameters(
+                lc = call.parameters[LANGUAGE_PARAM_KEY]
+            )
+            call.respond(
+                productsView(params, path, resolver, contentLanguage)
+            )
+        }
+    }
+}
+
+private fun productsView(
+    params: UrlParameters,
+    path: String,
+    resolver: DependencyResolver,
+    contentLanguage: List<Locale.LanguageRange>
+): ThymeleafContent {
+    val validator =
+        RoutingValidator(resolver)
+
+    if (!validator.isLanguageCodeValid(params.languageCode)) {
+        return errorPage(
+            "invalid_route_parameter",
+            "invalid_route_parameter_message",
+            HttpStatusCode.NotFound,
+            contentLanguage
+        )
+    }
+
+    val model = FetchProductViewData(resolver.productCatalog)
+    val languageName = getLanguageName(params.languageCode, resolver)
+
+    return ThymeleafContent(
+        template = "products",
+        model = mapOf(
+            "productList" to model.getListViewData(path),
+            "languagesNavTitle" to languageName,
+            "languagesNavUrl" to "/$GL_ROUTE",
+            "fileTypesNavUrl" to "#"
+        ),
+        locale = getPreferredLocale(contentLanguage, "products")
+    )
+}
