+++ src/main/kotlin/org/bibletranslationtools/fetcher/web/controllers/utils/ConstantValues.kt
@@ -1,0 +1,6 @@
+package org.bibletranslationtools.fetcher.web.controllers.utils
+
+const val GL_ROUTE = "gl"
+const val LANGUAGE_PARAM_KEY = "languageCode"
+const val PRODUCT_PARAM_KEY = "productSlug"
+const val BOOK_PARAM_KEY = "bookSlug"
